requirements.txt
# Minimum Python requirements
Python>=3.7

# Serial communication
pyserial>=3.5

# Data and date handling
python-dateutil>=2.8.2
pandas>=1.5.0

# Web interface
streamlit>=1.24.0

# AI and processing
langchain>=0.0.200
openai>=0.27.0

# JSON formatting (already included in Python standard library, listed for reference)
json

# Types (for type hints)
typing-extensions>=4.5.0

# ... other dependencies ...

# Garmin Connect API
garminconnect>=0.1.47

# Environment variables
python-dotenv>=0.19.0